+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * [HexaPDF::Layout::TableBox] for rendering tables
 * [HexaPDF::Content::GraphicObject::EndpointArc#max_curves] for setting the
   approximation accuracy
+* [HexaPDF::Importer::copy] for completely copying (including referenced
+  indirect objects) a single PDF object (which may be from the same document)
 
 ### Changed